+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## V2.1.0 - 2020-7-5
+
+### Added
+
+- debounce all shade changes as a group, instead of by shade id
+
+- this allows grouping all shades being set to the same position (with the same feature id) into a single set command to the bridge
+
 ## V2.0.5 - 2020-6-27
 
 ### Fixed